Move cell construction based on string to cells file
