+++ src/app/backend/cells.spec.ts
@@ -1,4 +1,14 @@
-import {CellType, determineCellTypeBasedOnValue, determineDirectionBasedOnValue, Direction, Enemy, Generator, Mover, Slider} from './cells';
+import {
+  CellType,
+  createCellInstanceFromString,
+  determineCellTypeBasedOnValue,
+  determineDirectionBasedOnValue,
+  Direction,
+  Enemy,
+  Generator, Immobile,
+  Mover, Push, Rotator,
+  Slider
+} from './cells';
 
 describe('backend/cells', () => {
   it('cells can be differentiated based on type', () => {
@@ -6,6 +16,14 @@ describe('backend/cells', () => {
     expect(mover instanceof Mover).toBeTrue();
     expect(mover instanceof Slider).toBeFalse();
     expect(mover.getCellType()).toEqual(CellType.MOVER);
+  });
+
+  it('cell instances can be created based on the string representation', () => {
+    const cells = [new Mover(Direction.LEFT), new Push(), new Slider(Direction.DOWN), new Rotator(), new Generator(Direction.UP),
+      new Immobile(), new Enemy()];
+    for (const cell of cells) {
+      expect(createCellInstanceFromString(cell.toString())).toEqual(cell);
+    }
   });
 
   it('toString() method gives cell type and optionally the direction.', () => {
